requirements.txt
alpaca-py==0.13.3
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
websockets==11.0.3
python-dotenv==1.0.0